Ostateczna wersja

+++ app/Http/Controllers/HomeController.php
@@ -12,6 +12,12 @@ use App\Models\Appointment;
 
 class HomeController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth',['only' => ['cancel_appoint','myappointment']]);
+    }
+
     public function redirect(){
         if(Auth::id()){
             if(Auth::user()->usertype=='0'){
@@ -36,41 +42,42 @@ class HomeController extends Controller
     }
 
     public function appointment(Request $request){
+        $this->validate($request, [
+            'name' => 'required|max:60|min:5|regex:/^[\pL\s\-]+$/u',
+            'number' => 'required|digits:9',
+            'email' => 'required|email|max:70|min:6',
+            'message' => 'max:200',
+            'date' => 'required|after_or_equal:today'
+        ]);
+
         $data = new appointment;
-        if(Auth::id()){
-            $data->user_id = Auth::user()->id;
-            $data->name = Auth::user()->name;
-            $data->email = Auth::user()->email;
-            $data->phone = Auth::user()->phone;
-        }else{
-            $data->name = $request->name;
-            $data->email = $request->email;
-            $data->phone = $request->number;
-        }
+        $data->name = $request->name;
+        $data->email = $request->email;
+        $data->phone = $request->number;
         $data->date = $request->date;
         $data->message = $request->message;
         $data->doctor = $request->doctor;
         $data->status = 'In progress';
-        
+     
         $data->save();
         return redirect()->back()->with('message','Appointment Successful');
     }
 
     public function myappointment(){
-        if(Auth::id()){
-            $user_mail = Auth::user()->email;
+        $user_mail = Auth::user()->email;
 
-            $appoint = appointment::where('email',$user_mail)->get();
-            return view('user.my_appointment',compact('appoint'));
-        }else{
-            return redirect()->back();
-        }
+        $appoint = appointment::where('email',$user_mail)->get();
+        return view('user.my_appointment',compact('appoint'));
+
         
     }
 
     public function cancel_appoint($id){
+        $id_user = Auth::user()->id;
         $data = appointment::find($id);
-        $data->delete();
-        return redirect()->back();
+        if($data->user_id == $id_user){
+            $data->delete();
+        }
+        return redirect()->back(); 
     }
 }
